api comments list for svist, swagger docs
