+++ src/Controller/Api/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Controller\Api;
 
 use App\Entity\User;
+use App\Entity\Svistyn;
 use App\Exception\JsonHttpException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -12,6 +13,7 @@ use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Knp\Component\Pager\PaginatorInterface;
 
 /**
  * Class UserController.
@@ -36,17 +38,24 @@ class UserController extends Controller
     private $validator;
 
     /**
+     * @var PaginatorInterface
+     */
+    private $paginator;
+
+    /**
      * UserController constructor.
      *
      * @param SerializerInterface          $serializer
      * @param UserPasswordEncoderInterface $passwordEncoder
      * @param ValidatorInterface           $validator
+     * @param PaginatorInterface           $paginator
      */
-    public function __construct(SerializerInterface $serializer, UserPasswordEncoderInterface $passwordEncoder, ValidatorInterface $validator)
+    public function __construct(SerializerInterface $serializer, UserPasswordEncoderInterface $passwordEncoder, ValidatorInterface $validator, PaginatorInterface $paginator)
     {
         $this->serializer = $serializer;
         $this->passwordEncoder = $passwordEncoder;
         $this->validator = $validator;
+        $this->paginator = $paginator;
     }
 
     /**
@@ -78,5 +87,32 @@ class UserController extends Controller
         }
 
         return $this->json(['user' => $user], Response::HTTP_OK);
+    }
+
+    /**
+     * @Route("/{id}/svist/page={page}", name="api_user_list_svist", methods={"GET"}, requirements={"id": "\d+", "page": "\d+"})
+     */
+    public function userListSvists($id, Request $request, string $page, $limit = 10)
+    {
+        $em = $this->getDoctrine()->getManager();
+        /** @var User $user */
+        $user = $em->getRepository(User::class)->findOneBy(['id' => $id]);
+
+        if (!$user) {
+            throw new JsonHttpException(Response::HTTP_NOT_FOUND, 'User not found');
+        }
+        $svists = $em->getRepository(Svistyn::class)->findBy(['user' => $user]);
+
+        if (!$svists) {
+            throw new JsonHttpException(Response::HTTP_NOT_FOUND, 'Svists not found');
+        }
+
+        return $this->json(
+          [
+            'svists' => $this->paginator->paginate(
+              $svists,
+              $request->query->getInt('page', $page), $limit),
+          ],
+          Response::HTTP_OK);
     }
 }
